Fixed leaks found by LeakCanary
